gym/requirements.txt
fastapi
uvicorn
jinja2
fpdf
google-generativeai
python-dotenv
